Remove dead code for multi-level nav links
The beautiful-hexo layout templates don't have
the necessary classes for beautiful-jekyll's
multi-level navigation links, and the necessary
partial layout templates are not included.

However, the CSS and JS for the multi-level nav
links from beautiful-jekyll is still there in
main.css and main.js. This should be removed,
to save bandwidth for site visitors and to make
maintenance easier.